Sign systemd boot EFI images for secure booting.
with efitools (`nix-shell -p efitools`)

and running with openssl in the shell environment:

    $ cd /home/grahamc/projects/grahamc/secure-boot
    $ uuidgen --random > GUID.txt
    $ openssl req -newkey rsa:4096 -nodes -keyout PK.key -new -x509 -sha256 -days 3650 -subj "/CN=my Platform Key/" -out PK.crt
    $ openssl x509 -outform DER -in PK.crt -out PK.cer
    $ cert-to-efi-sig-list -g "$(< GUID.txt)" PK.crt PK.esl
    $ sign-efi-sig-list -g "$(< GUID.txt)" -k PK.key -c PK.crt PK PK.esl PK.auth
    $ sign-efi-sig-list -g "$(< GUID.txt)" -c PK.crt -k PK.key PK /dev/null rm_PK.auth

    $ openssl req -newkey rsa:4096 -nodes -keyout KEK.key -new -x509 -sha256 -days 3650 -subj "/CN=my Key Exchange Key/" -out KEK.crt
    $ openssl x509 -outform DER -in KEK.crt -out KEK.cer
    $ cert-to-efi-sig-list -g "$(< GUID.txt)" KEK.crt KEK.esl
    $ sign-efi-sig-list -g "$(< GUID.txt)" -k PK.key -c PK.crt KEK KEK.esl KEK.auth

    $ openssl req -newkey rsa:4096 -nodes -keyout db.key -new -x509 -sha256 -days 3650 -subj "/CN=my Signature Database key/" -out db.crt
    $ openssl x509 -outform DER -in db.crt -out db.cer
    $ cert-to-efi-sig-list -g "$(< GUID.txt)" db.crt db.esl
    $ sign-efi-sig-list -g "$(< GUID.txt)" -k KEK.key -c KEK.crt db db.esl db.auth

and then loading the .auth files in my UEFI firmware, then setting in
my configuration.nix:

    boot = {
      loader = {
        systemd-boot = {
          enable = true;
          signed = true;
          signing-key = "/home/grahamc/projects/grahamc/secure-boot/db.key";
          signing-certificate = "/home/grahamc/projects/grahamc/secure-boot/db.crt";
        };
      };
    };

and running `nixos-rebuild boot` I was able to secure-boot my laptop.

Note: this code signs EVERY boot entry for NixOS, and the BOOT files
from systemd (even if they're already tampered with! since this code
doesn't replace it.) AND does not allow a graceful key phase out. If
the key is replaced in gen2, gen1's boot EFI is deleted and replaced
with a new version.

Also note this only works on x86-64 due to the assumptions in the
builder.py.

Finally, you can have much fewer generations: each generation on my
laptop takes 20MB in /boot.

Info taken from:

 - https://wiki.archlinux.org/index.php/Secure_Boot#Using_your_own_keys
 - https://systemd.io/BOOT_LOADER_SPECIFICATION#type-2-efi-unified-kernel-images
 - https://www.freedesktop.org/software/systemd/man/os-release.html